+++ lib/src/typed/typed_searchable_dropdown.dart
@@ -39,6 +39,10 @@ class TypedSearchableDropdown<T> extends StatelessWidget {
     this.errorBuilder,
     this.semanticsLabel,
     this.minSearchLength = 0,
+    this.onLoadMore,
+    this.hasMore = false,
+    this.isLoadingMore = false,
+    this.focusNode,
   });
 
   final String hintText;
@@ -68,6 +72,10 @@ class TypedSearchableDropdown<T> extends StatelessWidget {
   final DropdownErrorBuilder? errorBuilder;
   final String? semanticsLabel;
   final int minSearchLength;
+  final VoidCallback? onLoadMore;
+  final bool hasMore;
+  final bool isLoadingMore;
+  final FocusNode? focusNode;
 
   @override
   Widget build(BuildContext context) {
@@ -107,6 +115,10 @@ class TypedSearchableDropdown<T> extends StatelessWidget {
       errorBuilder: errorBuilder,
       semanticsLabel: semanticsLabel,
       minSearchLength: minSearchLength,
+      onLoadMore: onLoadMore,
+      hasMore: hasMore,
+      isLoadingMore: isLoadingMore,
+      focusNode: focusNode,
     );
   }
 }